rename dualsensex to dsx hud
