+++ Race Element.HUD.Common/Overlays/Pitwall/DSX/DsxJob.cs
@@ -1,15 +1,16 @@
 ﻿using RaceElement.Core.Jobs.Loop;
+using RaceElement.Data.Common;
 using System;
-using static RaceElement.HUD.ACC.Overlays.Pitwall.OverlayDualSenseX.DualSenseXResources;
+using static RaceElement.HUD.Common.Overlays.Pitwall.DSX.DsxResources;
 
-namespace RaceElement.HUD.ACC.Overlays.Pitwall.OverlayDualSenseX;
+namespace RaceElement.HUD.Common.Overlays.Pitwall.DSX;
 
-internal sealed class DualSenseXJob(DualSenseXOverlay overlay) : AbstractLoopJob
+internal sealed class DsxJob(DsxOverlay overlay) : AbstractLoopJob
 {
     public sealed override void RunAction()
     {
-        if (!overlay.ShouldRender())
-            return;
+        //if (!overlay.ShouldRender())
+        //    return;
 
         if (overlay._client == null)
         {
@@ -20,11 +21,11 @@ internal sealed class DualSenseXJob(DualSenseXOverlay overlay) : AbstractLoopJob
             }
             catch (Exception)
             {
-               // let's not cause an app crash, shall we?
+                // let's not cause an app crash, shall we?
             }
         }
 
-        Packet tcPacket = TriggerHaptics.HandleAcceleration(ref overlay.pagePhysics, overlay._config.ThrottleHaptics);
+        Packet tcPacket = TriggerHaptics.HandleAcceleration(overlay._config, overlay.GameWhenStarted);
         if (tcPacket != null)
         {
             overlay.Send(tcPacket);
@@ -32,7 +33,7 @@ internal sealed class DualSenseXJob(DualSenseXOverlay overlay) : AbstractLoopJob
             //HandleResponse(response);
         }
 
-        Packet absPacket = TriggerHaptics.HandleBraking(ref overlay.pagePhysics, overlay._config.BrakeHaptics);
+        Packet absPacket = TriggerHaptics.HandleBraking(overlay._config, overlay.GameWhenStarted);
         if (absPacket != null)
         {
             overlay.Send(absPacket);
